Cleanup: Notes and some Core Elements (#27)
* Moving example game state from core to readme

* Use dynamic move-to-cell function in roll handler

Replaces the hardcoded move-to-cell function call with the dynamic function from the game state. This allows for more flexible and customizable movement logic per game instance.

* Refactor initial game state construction

Simplifies the creation of the initial game state by removing an unnecessary intermediate binding and directly returning the state map. This improves code readability and maintainability.

* Remove progress and remaining logic sections from README

Deleted the detailed 'Current Progress' and 'Remaining Logic' sections from the README to streamline documentation and remove outdated or unnecessary project status information.

* Refactor game completion logic in core.cljc

Simplified the end-of-game condition to check for fewer than two active players instead of separate cases for one or zero. Removed redundant logging and comments for unreachable states.

* Comments

* Using goog.string.format to replace the cljs version of format/printf